jurnal

+++ src/main.book/doc/controller.js
@@ -57,44 +57,46 @@ exports.Controller = class {
         );
       });
       doc.childs = await DocMainBookDB.createDocChild(create_childs, client);
-      const endDoc = await DocMainBookDB.getDoc([region_id, budjet.id], year, month, 'end');
-      if (!endDoc.length) {
-        const end_doc = await DocMainBookDB.createDoc([
-          user_id,
-          budjet_id,
-          'end',
-          month,
-          year,
-          tashkentTime(),
-          tashkentTime()
-        ], client)
-        const create_childs = []
-        childs.forEach(item => {
-          create_childs.push(
-            item.spravochnik_main_book_schet_id,
-            end_doc.id,
-            item.debet_sum,
-            item.kredit_sum,
+      if (type_document !== 'end') {
+        const endDoc = await DocMainBookDB.getDoc([region_id, budjet.id], year, month, 'end');
+        if (!endDoc.length) {
+          const end_doc = await DocMainBookDB.createDoc([
+            user_id,
+            budjet_id,
+            'end',
+            month,
+            year,
             tashkentTime(),
             tashkentTime()
-          );
-        });
-        end_doc.childs = await DocMainBookDB.createDocChild(create_childs, client);
-      } else {
-        await DocMainBookDB.deleteDocChilds([endDoc[0].id], client);
-        const create_childs = []
-        const end_childs = await DocMainBookDB.getOperatsiiSum([region_id, year, month, budjet_id], client);
-        for (let child of end_childs) {
-          create_childs.push(
-            child.spravochnik_main_book_schet_id,
-            endDoc[0].id,
-            child.debet_sum,
-            child.kredit_sum,
-            tashkentTime(),
-            tashkentTime()
-          );
+          ], client)
+          const create_childs = []
+          childs.forEach(item => {
+            create_childs.push(
+              item.spravochnik_main_book_schet_id,
+              end_doc.id,
+              item.debet_sum,
+              item.kredit_sum,
+              tashkentTime(),
+              tashkentTime()
+            );
+          });
+          end_doc.childs = await DocMainBookDB.createDocChild(create_childs, client);
+        } else {
+          await DocMainBookDB.deleteDocChilds([endDoc[0].id], client);
+          const create_childs = []
+          const end_childs = await DocMainBookDB.getOperatsiiSum([region_id, year, month, budjet_id], client);
+          for (let child of end_childs) {
+            create_childs.push(
+              child.spravochnik_main_book_schet_id,
+              endDoc[0].id,
+              child.debet_sum,
+              child.kredit_sum,
+              tashkentTime(),
+              tashkentTime()
+            );
+          }
+          endDoc[0].childs = await DocMainBookDB.createDocChild(create_childs, client);
         }
-        endDoc[0].childs = await DocMainBookDB.createDocChild(create_childs, client);
       }
       return doc;
     })
@@ -201,44 +203,46 @@ exports.Controller = class {
         );
       });
       doc.childs = await DocMainBookDB.createDocChild(create_childs, client);
-      const endDoc = await DocMainBookDB.getDoc([region_id, budjet.id], year, month, 'end');
-      if (!endDoc.length) {
-        const end_doc = await DocMainBookDB.createDoc([
-          user_id,
-          budjet_id,
-          'end',
-          month,
-          year,
-          tashkentTime(),
-          tashkentTime()
-        ], client)
-        const create_childs = []
-        childs.forEach(item => {
-          create_childs.push(
-            item.spravochnik_main_book_schet_id,
-            end_doc.id,
-            item.debet_sum,
-            item.kredit_sum,
+      if (type_document !== 'end') {
+        const endDoc = await DocMainBookDB.getDoc([region_id, budjet.id], year, month, 'end');
+        if (!endDoc.length) {
+          const end_doc = await DocMainBookDB.createDoc([
+            user_id,
+            budjet_id,
+            'end',
+            month,
+            year,
             tashkentTime(),
             tashkentTime()
-          );
-        });
-        end_doc.childs = await DocMainBookDB.createDocChild(create_childs, client);
-      } else {
-        await DocMainBookDB.deleteDocChilds([endDoc[0].id], client);
-        const create_childs = []
-        const end_childs = await DocMainBookDB.getOperatsiiSum([region_id, year, month, budjet_id], client);
-        for (let child of end_childs) {
-          create_childs.push(
-            child.spravochnik_main_book_schet_id,
-            endDoc[0].id,
-            child.debet_sum,
-            child.kredit_sum,
-            tashkentTime(),
-            tashkentTime()
-          );
+          ], client)
+          const create_childs = []
+          childs.forEach(item => {
+            create_childs.push(
+              item.spravochnik_main_book_schet_id,
+              end_doc.id,
+              item.debet_sum,
+              item.kredit_sum,
+              tashkentTime(),
+              tashkentTime()
+            );
+          });
+          end_doc.childs = await DocMainBookDB.createDocChild(create_childs, client);
+        } else {
+          await DocMainBookDB.deleteDocChilds([endDoc[0].id], client);
+          const create_childs = []
+          const end_childs = await DocMainBookDB.getOperatsiiSum([region_id, year, month, budjet_id], client);
+          for (let child of end_childs) {
+            create_childs.push(
+              child.spravochnik_main_book_schet_id,
+              endDoc[0].id,
+              child.debet_sum,
+              child.kredit_sum,
+              tashkentTime(),
+              tashkentTime()
+            );
+          }
+          endDoc[0].childs = await DocMainBookDB.createDocChild(create_childs, client);
         }
-        endDoc[0].childs = await DocMainBookDB.createDocChild(create_childs, client);
       }
       return doc;
     })
